Moved List accessors to a separate file
