+++ packages/realm/src/List.ts
@@ -19,12 +19,12 @@
 import { assert } from "./assert";
 import { binding } from "../binding";
 import { COLLECTION_ACCESSOR as ACCESSOR } from "./Collection";
-import { Dictionary, createDictionaryAccessor, insertIntoDictionaryOfMixed, isJsOrRealmDictionary } from "./Dictionary";
 import { AssertionError, IllegalConstructorError } from "./errors";
-import { OrderedCollection, createDefaultGetter } from "./OrderedCollection";
+import { OrderedCollection } from "./OrderedCollection";
 import type { ObjectSchema } from "./schema";
 import type { Realm } from "./Realm";
 import { type TypeHelpers, toItemType } from "./TypeHelpers";
+import type { ListAccessor } from "./collection-accessors/List";
 
 type PartiallyWriteableArray<T> = Pick<Array<T>, "pop" | "push" | "shift" | "unshift" | "splice">;
 
@@ -309,135 +309,3 @@ export class List<T = unknown>
     this.internal.swap(index1, index2);
   }
 }
-
-/**
- * Accessor for getting, setting, and inserting items in the binding collection.
- * @internal
- */
-export type ListAccessor<T = unknown> = {
-  get: (list: binding.List, index: number) => T;
-  set: (list: binding.List, index: number, value: T) => void;
-  insert: (list: binding.List, index: number, value: T) => void;
-};
-
-type ListAccessorFactoryOptions<T> = {
-  realm: Realm;
-  typeHelpers: TypeHelpers<T>;
-  itemType: binding.PropertyType;
-  isEmbedded?: boolean;
-};
-
-/** @internal */
-export function createListAccessor<T>(options: ListAccessorFactoryOptions<T>): ListAccessor<T> {
-  return options.itemType === binding.PropertyType.Mixed
-    ? createListAccessorForMixed<T>(options)
-    : createListAccessorForKnownType<T>(options);
-}
-
-function createListAccessorForMixed<T>({
-  realm,
-  typeHelpers,
-}: Pick<ListAccessorFactoryOptions<T>, "realm" | "typeHelpers">): ListAccessor<T> {
-  const { toBinding } = typeHelpers;
-  return {
-    get(list, index) {
-      const value = list.getAny(index);
-      switch (value) {
-        case binding.ListSentinel: {
-          const accessor = createListAccessor<T>({ realm, typeHelpers, itemType: binding.PropertyType.Mixed });
-          return new List<T>(realm, list.getList(index), accessor, typeHelpers) as T;
-        }
-        case binding.DictionarySentinel: {
-          const accessor = createDictionaryAccessor<T>({ realm, typeHelpers, itemType: binding.PropertyType.Mixed });
-          return new Dictionary<T>(realm, list.getDictionary(index), accessor, typeHelpers) as T;
-        }
-        default:
-          return typeHelpers.fromBinding(value);
-      }
-    },
-    set(list, index, value) {
-      assert.inTransaction(realm);
-
-      if (isJsOrRealmList(value)) {
-        list.setCollection(index, binding.CollectionType.List);
-        insertIntoListOfMixed(value, list.getList(index), toBinding);
-      } else if (isJsOrRealmDictionary(value)) {
-        list.setCollection(index, binding.CollectionType.Dictionary);
-        insertIntoDictionaryOfMixed(value, list.getDictionary(index), toBinding);
-      } else {
-        list.setAny(index, toBinding(value));
-      }
-    },
-    insert(list, index, value) {
-      assert.inTransaction(realm);
-
-      if (isJsOrRealmList(value)) {
-        list.insertCollection(index, binding.CollectionType.List);
-        insertIntoListOfMixed(value, list.getList(index), toBinding);
-      } else if (isJsOrRealmDictionary(value)) {
-        list.insertCollection(index, binding.CollectionType.Dictionary);
-        insertIntoDictionaryOfMixed(value, list.getDictionary(index), toBinding);
-      } else {
-        list.insertAny(index, toBinding(value));
-      }
-    },
-  };
-}
-
-function createListAccessorForKnownType<T>({
-  realm,
-  typeHelpers,
-  itemType,
-  isEmbedded,
-}: Omit<ListAccessorFactoryOptions<T>, "isMixed">): ListAccessor<T> {
-  const { fromBinding, toBinding } = typeHelpers;
-  return {
-    get: createDefaultGetter({ fromBinding, itemType }),
-    set(list, index, value) {
-      assert.inTransaction(realm);
-      list.setAny(
-        index,
-        toBinding(value, isEmbedded ? { createObj: () => [list.setEmbedded(index), true] } : undefined),
-      );
-    },
-    insert(list, index, value) {
-      assert.inTransaction(realm);
-      if (isEmbedded) {
-        // Simply transforming to binding will insert the embedded object
-        toBinding(value, { createObj: () => [list.insertEmbedded(index), true] });
-      } else {
-        list.insertAny(index, toBinding(value));
-      }
-    },
-  };
-}
-
-/** @internal */
-export function insertIntoListOfMixed(
-  list: List | unknown[],
-  internal: binding.List,
-  toBinding: TypeHelpers["toBinding"],
-) {
-  // TODO: Solve the "removeAll()" case for self-assignment (https://github.com/realm/realm-core/issues/7422).
-  internal.removeAll();
-
-  for (const [index, item] of list.entries()) {
-    if (isJsOrRealmList(item)) {
-      internal.insertCollection(index, binding.CollectionType.List);
-      insertIntoListOfMixed(item, internal.getList(index), toBinding);
-    } else if (isJsOrRealmDictionary(item)) {
-      internal.insertCollection(index, binding.CollectionType.Dictionary);
-      insertIntoDictionaryOfMixed(item, internal.getDictionary(index), toBinding);
-    } else {
-      internal.insertAny(index, toBinding(item));
-    }
-  }
-}
-
-/** @internal */
-export function isJsOrRealmList(value: unknown): value is List | unknown[] {
-  return Array.isArray(value) || value instanceof List;
-}
-
-/* eslint-disable-next-line @typescript-eslint/no-explicit-any -- We define these once to avoid using "any" through the code */
-export type AnyList = List<any>;
